+++ Kernel/Language/ru.pm
@@ -9,6 +9,7 @@
 # Copyright (C) 2010-2011 Eugene Kungurov <ekungurov83 at ya.ru>
 # Copyright (C) 2010 Sergey Romanov <romanov_s at mail.ru>
 # Copyright (C) 2012-2013 Vadim Goncharov <vgoncharov at mail.ru>
+# Copyright (C) 2013 Alexey Gluhov <glalexnn at yandex.ru>
 # --
 # This software comes with ABSOLUTELY NO WARRANTY. For details, see
 # the enclosed file COPYING for license information (AGPL). If you
@@ -24,7 +25,7 @@ sub Data {
     my $Self = shift;
 
     # $$START$$
-    # Last translation file sync: 2013-05-28 11:37:26
+    # Last translation file sync: 2013-05-17 09:52:07
 
     # possible charsets
     $Self->{Charset} = ['utf-8', ];
@@ -263,9 +264,9 @@ sub Data {
         'Dispatching by email To: field.' => 'Перенаправление по заголовку To: электронного письма',
         'Dispatching by selected Queue.' => 'Перенаправление по выбранной очереди',
         'No entry found!' => 'Запись не найдена',
-        'Session invalid. Please log in again.' => '',
+        'Session invalid. Please log in again.' => 'Ошибка сессии. Пожалуйста авторизуйтесь вновь.',
         'Session has timed out. Please log in again.' => 'Сеанс завершен. Попробуйте войти заново.',
-        'Session limit reached! Please try again later.' => '',
+        'Session limit reached! Please try again later.' => 'Срок жизни сессии прошел. Пожалуйста попробуйте еще раз.',
         'No Permission!' => 'Нет доступа!',
         '(Click here to add)' => '(нажмите сюда чтобы добавить)',
         'Preview' => 'Предварительный просмотр',
@@ -356,7 +357,7 @@ sub Data {
         'more' => 'далее',
         'Collapse' => 'Свернуть',
         'Shown' => 'Показано',
-        'Shown customer users' => '',
+        'Shown customer users' => 'Отображенные клиенты',
         'News' => 'Новости',
         'Product News' => 'Новости о продукте',
         'OTRS News' => 'Новости OTRS',
@@ -697,7 +698,6 @@ sub Data {
         'Show the ticket history' => 'Показать историю заявки',
         'Print this ticket' => 'Печать этой заявки',
         'Print this article' => 'Печать этого сообщения',
-        'Split' => 'Разделить',
         'Split this article' => 'Разделить это сообщение',
         'Forward article via mail' => 'Переслать сообщение почтой',
         'Change the ticket priority' => 'Сменить приоритет заявки',
@@ -882,7 +882,7 @@ sub Data {
 
         # Template: AdminCustomerUser
         'Customer Management' => 'Управление клиентами',
-        'Back to search result' => '',
+        'Back to search result' => 'Назад к результатам поиска',
         'Add customer' => 'Добавить клиента',
         'Select' => 'Выбор',
         'Hint' => 'Подсказка',
@@ -891,7 +891,7 @@ sub Data {
         'Please enter a search term to look for customers.' => 'Введите запрос для поиска клиентов.',
         'Last Login' => 'Последний вход',
         'Login as' => 'Зайти данным пользователем',
-        'Switch to customer' => '',
+        'Switch to customer' => 'Переключится на клиента',
         'Add Customer' => 'Добавить клиента',
         'Edit Customer' => 'Редактировать клиента',
         'This field is required and needs to be a valid email address.' =>
@@ -979,8 +979,8 @@ sub Data {
             'Это порядок, в котором поле будет показываться среди других полей на тех экранах, где оно активно.',
         'Field type' => 'Тип поля',
         'Object type' => 'Тип объекта',
-        'Internal field' => '',
-        'This field is protected and can\'t be deleted.' => '',
+        'Internal field' => 'Внутреннее поле',
+        'This field is protected and can\'t be deleted.' => 'Это поле защищено и не может быть удалено.',
         'Field Settings' => ' - настройки поля',
         'Default value' => 'Значение по умолчанию',
         'This is the default value for this field.' => 'Это значение по умолчанию для данного поля.',
@@ -1115,7 +1115,7 @@ sub Data {
         'Archive selected tickets' => 'Архивировать выбранные заявки',
         'Add Note' => 'Добавить заметку',
         'Time units' => 'Единицы времени',
-        '(work units)' => '',
+        ' (work units)' => ' (рабочие единицы)',
         'Ticket Commands' => 'Команды по заявке',
         'Send agent/customer notifications on changes' => 'Отправлять уведомление агенту при изменениях',
         'CMD' => 'Команда',
@@ -1140,7 +1140,7 @@ sub Data {
         'GenericInterface Debugger for Web Service %s' => '',
         'Web Services' => 'Вебсервисы',
         'Debugger' => 'Отладчик',
-        'Go back to web service' => '',
+        'Go back to web service' => 'Перейти назад к веб сервису',
         'Clear' => 'Очистить',
         'Do you really want to clear the debug log of this web service?' =>
             '',
@@ -1469,7 +1469,7 @@ sub Data {
         'Install Package' => 'Установить пакет',
         'Update repository information' => 'Обновить информацию репозитория',
         'Did not find a required feature? OTRS Group provides their service contract customers with exclusive Add-Ons:' =>
-            '',
+            'Не нашли требующуюся функцию? Клиентам заключившим сервисный контракт, OTRS Group предлагает набор расширений(Add-One)',
         'Online Repository' => 'Онлайновый репозиторий',
         'Vendor' => 'Изготовитель',
         'Module documentation' => 'Документация модуля',
@@ -1606,12 +1606,12 @@ sub Data {
         'ArticleType' => '',
         'Display' => '',
         'Edit Field Details' => '',
-        'Customer interface does not support internal article types.' => '',
+        'Customer interface does not support internal article types.' => 'Клиентский интерфейс не поддерживает внутренний тип заметки',
 
         # Template: AdminProcessManagementPath
         'Path' => '',
-        'Edit this transition' => '',
-        'Transition Actions' => '',
+        'Edit this transition' => 'Редактировать это преобразование',
+        'Transition Actions' => 'Действия преобразований',
         'You can assign Transition Actions to this Transition by dragging the elements with the mouse from the left list to the right list.' =>
             '',
         'Filter available Transition Actions' => '',
@@ -2362,7 +2362,7 @@ sub Data {
         'Fulltext' => 'Полнотекстовый',
         'Remove' => 'Удалить',
         'Searches in the attributes From, To, Cc, Subject and the article body, overriding other attributes with the same name.' =>
-            '',
+            'Искать в свойствах От, Для, Копия, Тема и тело сообщения, перекрывая другие атрибуты с тем же именем.',
         'Customer User Login' => 'Логин клиента',
         'Created in Queue' => 'Создана в очереди',
         'Lock state' => 'Состояние блокировки',
@@ -2391,14 +2391,14 @@ sub Data {
         'Article Type' => 'Тип сообщения',
         'Sender Type' => 'Тип отправителя',
         'Save filter settings as default' => 'Сохранить условия фильтра для показа по умолчанию',
-        'Archive' => '',
-        'This ticket is archived.' => '',
+        'Archive' => 'Архив',
+        'This ticket is archived.' => 'Заявка перемещена в архив.',
         'Linked Objects' => 'Связанные объекты',
         'Article(s)' => 'сообщений',
         'Change Queue' => 'Сменить очередь',
         'There are no dialogs available at this point in the process.' =>
-            '',
-        'This item has no articles yet.' => '',
+            'Нет диалогов доступных в этой части процесса',
+        'This item has no articles yet.' => 'Этот элемент пока не имеет заметок.',
         'Article Filter' => 'Фильтр сообщений',
         'Add Filter' => 'Добавить фильтр',
         'Set' => 'Установить',
@@ -2515,10 +2515,10 @@ sub Data {
         'Search Results for' => 'Результаты поиска для',
 
         # Template: CustomerTicketZoom
-        'Show  article' => '',
+        'Show  article' => 'Показать сообщение',
         'Expand article' => 'Развернуть сообщение',
-        'Information' => '',
-        'Next Steps' => '',
+        'Information' => 'Сведения',
+        'Next Steps' => 'Далее',
         'Reply' => 'Ответить',
 
         # Template: CustomerWarning
@@ -2568,9 +2568,9 @@ sub Data {
         # Template: HTMLHeadBlockEvents
 
         # Template: Header
-        'Fulltext search' => '',
-        'CustomerID Search' => '',
-        'CustomerUser Search' => '',
+        'Fulltext search' => 'Полнотекстовый поиск',
+        'CustomerID Search' => 'Поиск по ID клиента',
+        'CustomerUser Search' => 'Поиск по логину',
         'You are logged in as' => 'Вы вошли как',
 
         # Template: HeaderSmall
@@ -2649,7 +2649,7 @@ sub Data {
 
         # Template: InstallerRegistration
         'Organization' => 'Организация',
-        'Position' => '',
+        'Position' => 'Сфера деятельности',
         'Complete registration and continue' => 'Завершить регистрацию и продолжить',
         'Please fill in all fields marked as mandatory.' => 'Пожалуйста, заполните все поля, отмеченные как обязательные.',
 
@@ -2674,7 +2674,7 @@ sub Data {
             'Вручную вводимые адреса электронной почты будут проверяться на MX-записи в DNS. Не используйте эту опцию, если ваш DNS-сервер медленный или не разрешает публичные адреса.',
 
         # Template: LinkObject
-        'Object#' => '',
+        'Object#' => 'Объект#',
         'Add links' => 'Добавить связи',
         'Delete links' => 'Удалить связи',
 
@@ -2702,7 +2702,7 @@ sub Data {
         # Template: PictureUpload
         'Need FormID!' => 'Требуется FormID!',
         'No file found!' => 'Файл не найден!',
-        'The file is not an image that can be shown inline!' => '',
+        'The file is not an image that can be shown inline!' => 'Этот файл не может быть отображен как часть текста!',
 
         # Template: PrintFooter
         'URL' => 'URL',
@@ -2844,8 +2844,6 @@ sub Data {
         'Automatically sets the responsible of a ticket (if it is not set yet) after the first owner update.' =>
             '',
         'Balanced white skin by Felix Niklas.' => '',
-        'Basic fulltext index settings. Execute "bin/otrs.RebuildFulltextIndex.pl" in order to generate a new index.' =>
-            '',
         'Blocks all the incoming emails that do not have a valid ticket number in subject with From: @example.com address.' =>
             '',
         'Builds an article index right after the article\'s creation.' =>
@@ -2880,6 +2878,8 @@ sub Data {
         'Configure Processes.' => '',
         'Configure your own log text for PGP.' => '',
         'Configures a default TicketDynmicField setting. "Name" defines the dynamic field which should be used, "Value" is the data that will be set, and "Event" defines the trigger event. Please check the developer manual (http://doc.otrs.org/), chapter "Ticket Event Module".' =>
+            '',
+        'Configures the full-text index. Execute "bin/otrs.RebuildFulltextIndex.pl" in order to generate a new index.' =>
             '',
         'Controls if customers have the ability to sort their tickets.' =>
             '',
@@ -3664,7 +3664,6 @@ sub Data {
         'Frontend module registration for the agent interface.' => '',
         'Frontend module registration for the customer interface.' => '',
         'Frontend theme' => 'Тема интерфеса',
-        'Fulltext index regex filters to remove parts of the text.' => '',
         'GenericAgent' => 'Планировщик задач',
         'GenericInterface Debugger GUI' => '',
         'GenericInterface Invoker GUI' => '',
@@ -4474,7 +4473,6 @@ sub Data {
             '',
         'Statistics' => 'Отчеты',
         'Status view' => 'Просмотр статусов',
-        'Stop words for fulltext index. These words will be removed.' => '',
         'Stores cookies after the browser has been closed.' => '',
         'Strips empty lines on the ticket preview in the queue view.' => '',
         'Textarea' => '',
@@ -4556,7 +4554,6 @@ sub Data {
         #
         # OBSOLETE ENTRIES FOR REFERENCE, DO NOT TRANSLATE!
         #
-        ' (work units)' => ' (рабочие единицы)',
         '%s Tickets affected! Do you really want to use this job?' => '%s заявок будет изменено! Выполнить это задание?',
         '(Checks MX recordes of used email addresses by composing an answer. Don\'t use CheckMXRecord if your OTRS machine is behinde a dial-up line $!)' =>
             'Проверять MX-записи домена, на который отправляется email при ответе. Не используйте эту возможность, если сервер с OTRS доступен по слабому каналу!',
@@ -5016,6 +5013,7 @@ sub Data {
         'Sort by' => 'Сортировка по',
         'Source' => 'Источник',
         'Spell Check' => 'Проверка орфографии',
+        'Split' => 'Разделить',
         'State Type' => 'Тип статуса',
         'Static-File' => 'Статический файл',
         'Stats-Area' => 'Статистика',